[CustomCNN class]
Created subclass of BaseFeaturesExtractor for using Keras pre-trained
CNNs to process image data. The default model used is Xception.

I've tested this with some 100,000 timesteps and the PPO agent seems to
learn but I'm not sure the best way to confirm this. I've seen average
episode lengths and rewards increase but I think the number of timesteps
may be too little to learn anything significant. I'll try visualizing
the behavior to see if anything qualitative appears.

Also - given that this is using a TF model behind the scenes, are the
weights in the final trainable layers getting updated? That is, is
backpropagation happening? This is a concern I have around the cross-use
between PyTorch and TensorFlow.