[Custom CNN] Fixed default model handling
There was a bug where the ResNet50 model/weights wouldn't get loaded
when no model was specified to the PPO agent. This should be fixed now.

Also, in the script I've commented out the section that uses the vision
transformer. In the notebook I've made a note about OOM issues using the
vision transformer.